fix: typed runtime.generate()
